bots: don't use qstat to kill jobs if the queue is 'none'

+++ Firebot/run_firebot.sh
@@ -161,10 +161,12 @@ if [ "$KILL_FIREBOT" == "1" ]; then
     kill -9 $(LIST_DESCENDANTS $PID)
     echo "killing firebot (PID=$PID)"
     kill -9 $PID
-    JOBIDS=`qstat -a | grep FB_ | awk -v user="$USER" '{if($2==user){print $1}}'`
-    if [ "$JOBIDS" != "" ]; then
-      echo killing firebot jobs with Id:$JOBIDS
-      qdel $JOBIDS
+    if [ "$QUEUE" != "none" ]; then
+      JOBIDS=`qstat -a | grep FB_ | awk -v user="$USER" '{if($2==user){print $1}}'`
+      if [ "$JOBIDS" != "" ]; then
+        echo killing firebot jobs with Id:$JOBIDS
+        qdel $JOBIDS
+      fi
     fi
     echo firebot process $PID killed
     if [ -e $firebot_pid ]; then
